chore: move assignment.ts to internals
