+++ src/internals/assignment.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  **/
 
-import { Experiment } from './api';
+import { Experiment } from '../api';
 
 export function randomAssignment(experiment: Experiment): boolean {
   const rand = Math.random() * 100;
